Fix data mismatch in two tables #1

+++ Base_de_datos/Tablas/CrearTablas.sql
@@ -454,7 +454,7 @@ CREATE TABLE dbo.AguaDePropiedad
 
     -- Otras columnas
     numeroMedidor INT NOT NULL,
-    consumoAcumulado float NOT NULL,
+    consumoAcumulado MONEY NOT NULL,
 
     -- Se establece la llave primaria
     CONSTRAINT PK_AguaDePropiedad PRIMARY KEY CLUSTERED (id),
